+++ ws_L_v.inmop_curloop/HC32F460_DDL_Rev3.3.0/projects/ev_hc32f460_lqfp100_v2/ws/foc.h
@@ -54,9 +54,9 @@ extern volatile float    g_foc_openloop_volt_v;  /* voltage amplitude (V), Keil 
 
 /* Current-loop run mode / observables (all volatile, Keil Watch editable) */
 extern volatile uint8_t  g_foc_mode;             /* 0=idle, 1=open-loop, 2=current-loop */
-extern volatile int8_t   g_foc_cur_sign;         /* current sign correction (+1/-1) */
-extern volatile int8_t   g_foc_enc_dir;         /* encoder direction for electrical angle (+1/-1), Watch tunable */
-extern volatile int8_t   g_foc_pi_off_180;     /* add 180deg to control angle to flip torque direction (0/1), Watch tunable */
+extern volatile int32_t  g_foc_cur_sign;        /* current sign correction (+1/-1), Watch tunable */
+extern volatile int32_t  g_foc_enc_dir;         /* encoder direction for electrical angle (+1/-1), Watch tunable */
+extern volatile int32_t  g_foc_pi_off_180;      /* add 180deg to control angle to flip torque direction (0/1), Watch tunable */
 extern volatile float    g_foc_iq_ref_cmd_ma;    /* Iq target (mA), Keil Watch editable */
 extern volatile float    g_foc_iq_ref_ma;        /* ramped Iq setpoint actually used (mA) */
 extern volatile float    g_foc_id_ma;            /* d-axis feedback (mA) */
